+++ client/app/components/admin/admin.component.js
@@ -6,13 +6,13 @@ import routes from './admin.routes';
 import UibTabsDirective from 'angular-ui-bootstrap/src/tabs';
 import AnnouncementManagerComponent from '../announcementmanager/announcementmanager.component';
 import ScheduleManagerComponent from '../schedulemanager/schedulemanager.component';
-import UserManagerComponent from '../usermanager/usermanager.component';
+import UserManagerModule from '../../modules/usermanager/usermanager.module';
 
 export class AdminComponent {
   /*@ngInject*/
 }
 
-export default angular.module('shyApp.admin', [ngRoute, AnnouncementManagerComponent, UibTabsDirective, ScheduleManagerComponent, UserManagerComponent])
+export default angular.module('shyApp.admin', [ngRoute, AnnouncementManagerComponent, UibTabsDirective, ScheduleManagerComponent, UserManagerModule])
   .config(routes)
   .component('admin', {
     template: require('./admin.pug'),