+++ client/app/components/admin/admin.routes.js
@@ -6,6 +6,9 @@ export default function routes($routeProvider) {
     url: '/admin',
     template: '<admin></admin>',
     authenticate: 'admin',
-    title: 'Schoolhouse Yoga Admin'
+    title: 'Schoolhouse Yoga Admin',
+    resolve: {
+      '': (TeachersService, ClassesService, LocationsService) => Promise.all([TeachersService.initialized, ClassesService.initialized, LocationsService.initialized])
+    }
   });
 }