+++ client/app/modules/shynet/shynet.module.js
@@ -4,11 +4,11 @@ import ngRoute from 'angular-route';
 import routes from './shynet.routes';
 import UibDatepickerPopupDirective from 'angular-ui-bootstrap/src/datepickerPopup/index-nocss.js';
 import UibModalDirective from 'angular-ui-bootstrap/src/modal/index-nocss.js';
-import UserManagerComponent from '../../components/usermanager/usermanager.component';
+import UserManagerModule from '../../modules/usermanager/usermanager.module';
 import { SHYnetComponent } from './shynet.component';
 import { AttendanceService } from './attendance.service';
 
-export default angular.module('shyApp.shynet', [ngRoute, UibDatepickerPopupDirective, UibModalDirective, UserManagerComponent])
+export default angular.module('shyApp.shynet', [ngRoute, UibDatepickerPopupDirective, UibModalDirective, UserManagerModule])
   .config(routes)
   .service('AttendanceService', AttendanceService)
   .component('shynet', {