+++ client/app/modules/usermanager/usereditor.controller.js
@@ -1,0 +1,73 @@
+'use strict';
+
+// Controller for modal dialog for editing users
+export class UserEditorController {
+  /*@ngInject*/
+  constructor($uibModalInstance, Auth, User, userSelectedForEditing) {
+    // Dependencies
+    this.$uibModalInstance = $uibModalInstance;
+    this.Auth = Auth;
+    this.User = User;
+    this.userSelectedForEditing = userSelectedForEditing;
+
+    // Initializations - not in $onInit since not it's own component
+    this.submitted = false;
+    this.errors = {};
+    this.user = {};
+    this.isAdmin = this.Auth.isAdminSync;
+    if(this.userSelectedForEditing) angular.copy(this.userSelectedForEditing, this.user);
+  }
+
+  // Reset server-side error status
+  clearServerError(form, fieldName) {
+    form[fieldName].$setValidity('server', true);
+  }
+
+  // Submit the modified user to the server
+  submitUser(form) {
+    this.submitted = true;
+    if(form.$valid) {
+      // Make a copy of this.user in case upsert fails
+      let upsertedUser = {};
+      angular.copy(this.user, upsertedUser);
+      this.User.upsert(upsertedUser)
+        .$promise
+        .then(user => { // only contains user._id
+          // Do not add the password and passwordConfirm to the array
+          Reflect.deleteProperty(upsertedUser, 'password'); // clear this out for security reasons
+          Reflect.deleteProperty(upsertedUser, 'passwordConfirm'); // ditto
+
+          // If a new user...
+          if(upsertedUser._id === 0) {
+            upsertedUser._id = user._id;
+          }
+
+          // Graft the edited user back the original
+          angular.extend(this.userSelectedForEditing, upsertedUser);
+          this.$uibModalInstance.close();
+          return null;
+        })
+        .catch(response => {
+          let err = response.data;
+          this.errors = {}; // reset to only the latest errors
+
+          // Update validity of form fields that match the server errors
+          if(err.name) {
+            for(let error of err.errors) {
+              form[error.path].$setValidity('server', false);
+              this.errors[error.path] = error.message;
+            }
+          }
+          return null;
+        });
+    }
+  }
+
+  // Cancel the user editor dialog
+  cancel() {
+    if(!this.userSelectedForEditing._id) {
+      this.userSelectedForEditing.shouldBeDeleted = true;
+    }
+    this.$uibModalInstance.close();
+  }
+}